refactor(omp): narrow completion spinner fix
