+++ src/main/pi/agent-status-extension-source.ts
@@ -250,7 +250,6 @@ export function getPiAgentStatusExtensionSource(kind: PiAgentKind = 'pi'): strin
     '    }, HOOK_POST_TIMEOUT_MS)',
     "    if (typeof timeout.unref === 'function') timeout.unref()",
     '  })',
-    '  let delivered = false',
     '  try {',
     '    const response = await Promise.race([',
     '      fetch(url, {',
@@ -264,15 +263,16 @@ export function getPiAgentStatusExtensionSource(kind: PiAgentKind = 'pi'): strin
     '      }),',
     '      timeoutPromise,',
     '    ])',
-    '    delivered = response.ok',
+    '    return response.ok',
     '  } catch {',
-    '    // Why: status delivery falls through to the WSL bridge or retry.',
+    '    // Why: status reporting must never fail the pi run just because Orca',
+    '    // is unavailable or the loopback request failed (e.g. Orca restart).',
+    '    if (!isWslRuntime()) return false',
+    '    postViaWindowsCurl(body, ompRuntime)',
+    '    return true',
     '  } finally {',
     '    if (timeout) clearTimeout(timeout)',
     '  }',
-    '  if (delivered) return true',
-    '  if (!isWslRuntime()) return false',
-    '  return postViaWindowsCurl(body, ompRuntime)',
     '}',
     '',
     ...getPiAgentStatusWslCurlSourceLines(),
