+++ src/main/pi/agent-status-wsl-curl-source.ts
@@ -43,11 +43,11 @@ export function getPiAgentStatusWslCurlSourceLines(): string[] {
     '}',
     '',
     '// Why: WSL loopback is not the Windows loopback, so use curl.exe on the host.',
-    'async function postViaWindowsCurl(body: string, ompRuntime: boolean): Promise<boolean> {',
+    'function postViaWindowsCurl(body: string, ompRuntime: boolean): void {',
     '  const curlPath = resolveWindowsCurlPath()',
     '  const windowsPort = process.env.ORCA_AGENT_HOOK_PORT',
     '  const windowsToken = process.env.ORCA_AGENT_HOOK_TOKEN',
-    '  if (!curlPath || !windowsPort || !windowsToken) return false',
+    '  if (!curlPath || !windowsPort || !windowsToken) return',
     '  // Why: a stale guest endpoint must fall back to current host coordinates.',
     '  const windowsUrl = `http://127.0.0.1:${windowsPort}${resolveHookPath(ompRuntime)}`',
     '  try {',
@@ -56,7 +56,6 @@ export function getPiAgentStatusWslCurlSourceLines(): string[] {
     '      curlPath,',
     '      [',
     "        '-sS',",
-    "        '-f',",
     '        // Why: detached delivery may take seconds under loaded WSL interop.',
     "        '--connect-timeout', '3',",
     "        '--max-time', '10',",
@@ -70,24 +69,11 @@ export function getPiAgentStatusWslCurlSourceLines(): string[] {
     '      ],',
     "      { stdio: ['pipe', 'ignore', 'ignore'] }",
     '    )',
-    '    return await new Promise<boolean>((resolve) => {',
-    '      let settled = false',
-    '      const finish = (delivered: boolean): void => {',
-    '        if (settled) return',
-    '        settled = true',
-    '        resolve(delivered)',
-    '      }',
-    "      child.on('error', () => finish(false))",
-    "      child.on('close', (code: number | null) => finish(code === 0))",
-    "      child.stdin.on('error', () => finish(false))",
-    '      try {',
-    '        child.stdin.end(body)',
-    '      } catch {',
-    '        finish(false)',
-    '      }',
-    '    })',
+    "    child.on('error', () => {})",
+    "    child.stdin.on('error', () => {})",
+    '    child.stdin.end(body)',
     '  } catch {',
-    '    return false',
+    '    // Why: status delivery must not surface inside the agent TUI.',
     '  }',
     '}',
     ''
